+++ src/components/companyProducts/AddProduct.js
@@ -142,7 +142,7 @@ function AddProduct(props) {
         />
         <TextField
           value={price}
-          onChange={(e) => setPrice(parseInt(e.target.value))}
+          onChange={(e) => setPrice(parseFloat(e.target.value))}
           autoFocus
           margin="normal"
           id="price"
